dcgan_face/requirements.txt
streamlit
tensorflow
Pillow
matplotlib
